backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
openai>=1.0.0
python-dotenv>=1.0.0
python-docx>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart==0.0.22
pypdf>=4.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
requests>=2.28.0
langdetect>=1.0.9